pyproject.toml
[project]
name = "realestate-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.14.1",
    "google-genai>=1.38.0",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
]

[tool.setuptools]
packages = ["real_estate_agent"]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]
